+++ grails-app/services/org/pih/warehouse/stockTransfer/StockTransferEventService.groovy
@@ -9,24 +9,18 @@
  **/
 package org.pih.warehouse.stockTransfer
 
-import grails.core.GrailsApplication
 import grails.gorm.transactions.Transactional
 import org.pih.warehouse.api.StockTransfer
-import org.pih.warehouse.api.StockTransferItem
 import org.pih.warehouse.core.ActivityCode
-import org.pih.warehouse.core.MailService
-import org.pih.warehouse.core.RoleType
-import org.pih.warehouse.core.UserService
 import org.pih.warehouse.order.Order
+import org.pih.warehouse.report.NotificationService
 import org.springframework.transaction.event.TransactionPhase
 import org.springframework.transaction.event.TransactionalEventListener
 
 @Transactional
 class StockTransferEventService {
 
-    GrailsApplication grailsApplication
-    UserService userService
-    MailService mailService
+    NotificationService notificationService
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT, fallbackExecution = true)
     void onStockTransfer(StockTransferEvent event) {
@@ -43,35 +37,7 @@ class StockTransferEventService {
             (it.originBinLocation?.supports(ActivityCode.ENABLE_STOCK_TRANSFER_NOTIFICATIONS)
                     || it.destinationBinLocation?.supports(ActivityCode.ENABLE_STOCK_TRANSFER_NOTIFICATIONS))
         }) {
-            sendStockTransferNotification(stockTransfer)
-        }
-    }
-
-    private void sendStockTransferNotification(StockTransfer stockTransfer) {
-        try {
-            def recipientList = userService.findUsersByRoleType(RoleType.ROLE_STOCK_TRANSFER_NOTIFICATION).collect {
-                it.email
-            }
-            if (recipientList) {
-                def g = grailsApplication.mainContext.getBean('org.grails.plugins.web.taglib.ApplicationTagLib')
-                // try to find StockTransferItem with destinationBinLocation that supports ENABLE_STOCK_TRANSFER_NOTIFICATIONS activity
-                List<StockTransferItem> stockTransferItemWithSupportedLocations = stockTransfer?.stockTransferItems?.findAll { StockTransferItem it ->
-                    (it.destinationBinLocation?.supports(ActivityCode.ENABLE_STOCK_TRANSFER_NOTIFICATIONS)
-                            || it.originBinLocation?.supports(ActivityCode.ENABLE_STOCK_TRANSFER_NOTIFICATIONS))
-                }
-                // unique names of origin/destination bin locations that support stock transfer notifications
-                List<String> uniqueBinNames = stockTransfer?.stockTransferItems
-                        ?.collectMany { StockTransferItem it -> [it.originBinLocation, it.destinationBinLocation] }
-                        ?.findAll { it?.supports(ActivityCode.ENABLE_STOCK_TRANSFER_NOTIFICATIONS) }
-                        ?.collect { it.name }
-                        ?.unique()
-                String subject = g.message(code: 'email.stockTransfer.message.subject')
-                def body = "${g.render(template: '/email/stockTransfer', model: [stockTransfer: stockTransfer, stockTransferItems: stockTransferItemWithSupportedLocations, uniqueBinNames: uniqueBinNames])}"
-                mailService.sendHtmlMail(subject, body.toString(), recipientList)
-            }
-        }
-        catch (Exception e) {
-            log.error("Error sending stock transfer notification email: " + e.message, e)
+            notificationService.sendStockTransferNotification(stockTransfer)
         }
     }
 }
